Crate pallet_crowdloan_claim

Crowdloan claim pallet

A pallet used for claiming reward payouts for crowdloan campaigns. This does not implement the rewarding strategy, that is the role of the [pallet-crowdloan-reward] pallet.

[Config]

[Call]

[Pallet]

Overview

This pallet is used to proces reward claims from contributors who locked tokens on the Polkadot/Kusama relay chain for participating in a crowdloan campaign for a parachain slot acquisition.

This Pallet is intimately bound to the [pallet-crowdloan-reward] pallet, where the rewarding strategy is implemented.

Terminology

For information on terms and concepts used in this pallet, please refer to the pallet’ specification document.

Goals

The aim of this pallet is to ensure that a contributor who’s claiming a reward is eligible for it and avoid potential attacks, such as, for instance, Denial of Service (DoS) or spams by requiring fee payment.

Usage

Interface

Supported Origins

Valid origin is an administrator or root.

Types

Events

Errors

Dispatchable Functions

Callable functions (or extrinsics), also considered as transactions, materialize the pallet contract. Here’s the callable functions implemented in this Pallet:

[claim_reward]

Public Functions

Genesis Configuration

Dependencies

As stated in the overview of this pallet, the latter relies on the [pallet-crowdloan-reward] pallet to implement the specific rewarding strategy of a parachain. A rewarding pallet must implement the [pallet-crowdloan-claim::traits::Reward] trait that is declared in this pallet. For this pallet to be able to interact with a reward pallet, it must be loosely coupled to the former using an associated type which includes the [pallet-crowdloan-claim::traits::Reward] trait.

References

Building a Custom Pallet. Retrieved April 5th, 2021.

Credits

The Centrifugians Tribe tribe@centrifuge.io

Re-exports

pub use pallet::*;

pub use crate::weights::WeightInfo;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_crowdloan_claim