Trait pallet_crowdloan_claim::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type PalletId: Get<PalletId>;
    type RelayChainAccountId: Debug + MaybeSerialize + MaybeSerializeDeserialize + Member + Ord + Parameter + Into<AccountId32>;
    type MaxProofLength: Get<u32>;
    type RewardMechanism: Reward<ParachainAccountId = Self::AccountId, ContributionAmount = Self::Balance, BlockNumber = Self::BlockNumber>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type WeightInfo: WeightInfo;
}

Crowdloan claim pallet’s configuration trait.

Associated types and constants are declared in this trait. If the pallet depends on other super-traits, the latter must be added to this trait, such as, in this case, [frame_system::Config] and [pallet_balances::Config] super-traits. Note that [frame_system::Config] must always be included.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Associated type for Event enum

type PalletId: Get<PalletId>

Constant configuration parameter to store the pallet identifier for the pallet.

The pallet identifier may be of the form PalletId(*b"cc/claim").

type RelayChainAccountId: Debug + MaybeSerialize + MaybeSerializeDeserialize + Member + Ord + Parameter + Into<AccountId32>

Contributor’s account identifier on the relay chain.

type MaxProofLength: Get<u32>

The maximum length (i.e. depth of the tree) we allow a proof to have. This mitigates DDoS attacks solely. We choose 30, which by a base 2 merkle-tree should be more than enough.

type RewardMechanism: Reward<ParachainAccountId = Self::AccountId, ContributionAmount = Self::Balance, BlockNumber = Self::BlockNumber>

The reward payout mechanism this claim pallet uses.

This associated type allows to implement a loosely-coupled regime between claiming and rewarding pallets.

type AdminOrigin: EnsureOrigin<Self::Origin>

Entity which is allowed to perform administrative transactions

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet