Type Alias pallet_crowdloan_claim::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn claim_reward( origin: OriginFor<T>, relaychain_account_id: T::RelayChainAccountId, parachain_account_id: <<T as Config>::RewardMechanism as Reward>::ParachainAccountId, identity_proof: MultiSignature, contribution_proof: Proof<T::Hash>, contribution: T::Balance ) -> DispatchResultWithPostInfo

Claim for a reward payout

pub fn initialize( origin: OriginFor<T>, contributions: <T as Config>::Hash, locked_at: T::BlockNumber, index: u32, lease_start: T::BlockNumber, lease_period: T::BlockNumber ) -> DispatchResultWithPostInfo

Initialize the claim pallet

This administrative function is used to transfer the list of contributors and their respective contributions, stored as a child trie root hash in the relay chain’s crowdloan pallet, to Contributions storage item. This transaction can only be called via a signed transactions. The contributions parameter contains the hash of the crowdloan pallet’s child trie root. It is later used for proving that a contributor effectively contributed to the crowdloan campaign, and that the amount of the contribution is correct as well.

pub fn set_lease_start( origin: OriginFor<T>, start: T::BlockNumber ) -> DispatchResultWithPostInfo

Set the start of the lease period.

pub fn set_lease_period( origin: OriginFor<T>, period: T::BlockNumber ) -> DispatchResultWithPostInfo

Set the lease period.

pub fn set_contributions_root( origin: OriginFor<T>, root: <T as Config>::Hash ) -> DispatchResultWithPostInfo

Set the root-hash of the relay-chain, we locked the relay-chain contributions at.

This root-hash MUST be the root-hash of the relay-chain at the block we locked at. This root-hash will be used to verify proofs of contribution.

pub fn set_locked_at( origin: OriginFor<T>, locked_at: T::BlockNumber ) -> DispatchResultWithPostInfo

Set the block of the relay at which we lock the contributions.

This means, that all generated proofs MUST generate the proof of their contribution at this block, as otherwise the root-hash we store here will not be found in the generated proof of the contributor, which will lead to a rejection of the proof.

pub fn set_crowdloan_trie_index( origin: OriginFor<T>, trie_index: u32 ) -> DispatchResultWithPostInfo

Set the index of the crowdloan.

This index comes from the relay-chain crowdloan pallet. More specifically, this index is used to derive the internal patricia key inside the child trie. The index is stored in the FundInfo of the relay chain crowdloan pallet.

impl<T: Config> Pallet<T>

pub fn contributions() -> Option<<T as Config>::Hash>

Root of hash of the relay chain at the time of initialization.

impl<T: Config> Pallet<T>

pub fn locked_at() -> Option<T::BlockNumber>

impl<T: Config> Pallet<T>

pub fn crowdloan_trie_index() -> Option<u32>

TrieIndex of the crowdloan campaign inside the relay-chain crowdloan pallet.

This is needed in order to build the correct keys for proof check.

impl<T: Config> Pallet<T>

pub fn processed_claims<KArg>(k: KArg) -> Option<bool>where KArg: EncodeLike<(T::RelayChainAccountId, u32)>,

A map containing the list of claims for reward payouts that were successfuly processed

impl<T: Config> Pallet<T>

pub fn curr_index() -> u32

impl<T: Config> Pallet<T>

pub fn prev_index() -> u32

impl<T: Config> Pallet<T>

pub fn lease_start() -> T::BlockNumber

impl<T: Config> Pallet<T>

pub fn lease_period() -> T::BlockNumber

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

Return the account identifier of the crowdloan claim pallet.

This actually does computation. If you need to keep using it, then make sure you cache the value and only call this once.